wrote the ws2812 driver

+++ include/stdperiph/timer/timer.h
@@ -22,9 +22,9 @@ uint16_t freqToCount(uint32_t target_freq, uint32_t prescale, uint32_t input_fre
 	return (uint16_t)(((1.0f / (float)target_freq) / ((float)prescale / (float)input_freq)) - 1.0f);
 }
 
-uint16_t periodToCount(float period, uint32_t prescale, uint32_t input_freq)
+uint16_t periodToCount(double period, uint32_t prescale, uint32_t input_freq)
 {
-	return (uint16_t)((period / ((float)prescale / (float)input_freq) ) - 1);
+	return (uint16_t)((period / ((double)prescale / (double)input_freq) ) - 1);
 }
 
 struct TimerHandler
